Move Subst to its own module.
