+++ src/Data/Complexity.hs
@@ -9,48 +9,19 @@ import qualified Control.Monad.Trans.Free as F
 import Control.Monad.Fresh
 import Control.Monad.Reader
 import Control.Monad.State
-import Data.Bifunctor (second)
 import Data.Env
 import Data.Expr
-import Data.Function (on)
 import Data.Functor.Foldable (Recursive(..), Fix(..))
-import qualified Data.List as List
 import Data.Maybe (fromMaybe)
 import Data.Name
-import Data.Semigroup (Semigroup(..))
 import qualified Data.Set as Set
+import Data.Subst
 import Data.Type
 
 data Complexity i
   = Constant i
   deriving (Eq, Ord, Read, Show)
 
-
-newtype Subst value = Subst { getSubst :: [(Name, value)] }
-  deriving (Eq, Foldable, Functor, Ord, Read, Show, Traversable)
-
-instance Binder value => Semigroup (Subst value) where
-  Subst s1 <> Subst s2 = Subst (List.unionBy ((==) `on` fst) (map (second (substitute (Subst s1))) s2) s1)
-
-instance Binder value => Monoid (Subst value) where
-  mempty = Subst []
-  mappend = (<>)
-
-substLookup :: Name -> Subst value -> Maybe value
-substLookup name = lookup name . getSubst
-
-substDelete :: Name -> Subst value -> Subst value
-substDelete name = Subst . filter ((/= name) . fst) . getSubst
-
-substSingleton :: Name -> value -> Subst value
-substSingleton name value = Subst [(name, value)]
-
-substExtend :: Binder value => Name -> value -> Subst value -> Subst value
-substExtend name value = (substSingleton name value <>)
-
-
-class Binder value where
-  substitute :: Subst value -> value -> value
 
 instance Binder (Free Type Error) where
   substitute = flip (cata (\ ty subst -> case ty of
